fix(stm32mp2): cast ST_CTX_VER_MAJOR for ST_CTX_VERSION
ST_CTX_VER_MAJOR (= 1U) is seen as an uint8_t. As it is shifted by 8, it
should then first be cast as an uint16_t when building ST_CTX_VERSION.

This corrects the MISRA_C_2012_Rule_12.2: The right hand operand of a
shift operator shall lie in the range zero to one less than the width
in bits of the essential type of the left hand operand

Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/541431
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Domain-Review: Patrick DELAUNAY <patrick.delaunay@foss.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>

+++ plat/st/stm32mp2/stm32mp2_context.c
@@ -31,7 +31,7 @@
 /* VERSION to check compatibility of pm backup data between bl2 and bl31 */
 #define ST_CTX_VER_MAJOR		1U
 #define ST_CTX_VER_MINOR		0U
-#define ST_CTX_VERSION			((ST_CTX_VER_MAJOR << 8U) | ST_CTX_VER_MINOR)
+#define ST_CTX_VERSION			(((uint16_t)ST_CTX_VER_MAJOR << 8U) | ST_CTX_VER_MINOR)
 #define ST_CTX_VER_GET_MAJOR(version)	(((version) & GENMASK_32(15, 8)) >> 8U)
 #define ST_CTX_VER_GET_MINOR(version)	((version) & GENMASK_32(7, 0))
